systemError: state refactor
