+++ webapp/app/appErrorsMiddleware.js
@@ -1,7 +1,7 @@
 import axios from 'axios'
 import * as R from 'ramda'
 
-import { appErrorCreate } from './appErrors/actions'
+import { appErrorCreate } from '../store/system/serviceError/actions'
 
 const createAxiosMiddleware = (axios) => ({ dispatch }) => {
   axios.interceptors.response.use(null, (error) => {
