+++ webapp/store/system/serviceError/reducer.js
@@ -1,8 +1,8 @@
-import { exportReducer } from '../../utils/reduxUtils'
+import { exportReducer } from '@webapp/utils/reduxUtils'
 
 import { appErrorCreate, appErrorDelete } from './actions'
 
-import * as ErrorsState from './appErrorsState'
+import * as ErrorsState from './state'
 
 const actionHandlers = {
   [appErrorCreate]: (state, { error }) => ErrorsState.assocAppError(error)(state),
